topluluk-backend/requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
PyJWT==2.9.0
sqlparse==0.5.3
channels==4.2.2
channels-redis==4.2.1
daphne==4.2.1
django-extensions==3.2.3
psycopg2-binary==2.9.9
ipython
pgvector
sentence-transformers